Moved table components into tables subdirectory
